[postgresql] Avoid unnecessary password lookup
When a password is defined for a role, the lookup of the password should
be completely avoided. The previous default construct was evaluated in
a way that would still access the password lookup, and potentially
generate the password, even if one was explicitly provided.

The result was correct, but since the password lookup can have
side effects, even in check-mode (it creates files), it should not
be run if it is not needed.